1.初步实现了添加到购物车功能，但是目前结算后仍无法修改库存 2.更正了语言问题，但是弹窗按钮仍是英语 3.更正了查找函数

+++ h/global.h
@@ -2,4 +2,4 @@
 #include "library.h"
 extern set<BookData> books;
 int my_atoi(const char* const &);
-int my_atof(const char* const &);
+double my_atof(const char* const &);
